+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - **Aprobar Matrícula**: nuevo tipo de matrícula posible, `IDO`.
 - Posibilidad de ordenar datos por columna en `ListaSolicitud` y `ListaMatriculas`.
+- **Nuevo Profesional**: posibilidad de firmar con canvas.
 
 
 # 1.1.1 (2018-01-22)